Refactor ingestion constants and update import paths
This commit removes the constants.ts file and updates import statements across various modules to reference the new core/constants module. Additionally, it modifies the types used in the establishRelations and insertBatch functions to utilize CodeModelRelation and CodeModel, respectively, enhancing type safety and code organization.
