+++ ingestion/src/core/constants.ts
